+++ api/core/v2/filter.pb.go
@@ -633,10 +633,7 @@ func (m *EventFilter) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthFilter
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthFilter
 			}
 			if (iNdEx + skippy) > l {