+++ api/core/v2/mutator.pb.go
@@ -682,10 +682,7 @@ func (m *Mutator) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthMutator
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthMutator
 			}
 			if (iNdEx + skippy) > l {