+++ api/core/v2/tessen.pb.go
@@ -348,10 +348,7 @@ func (m *TessenConfig) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthTessen
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthTessen
 			}
 			if (iNdEx + skippy) > l {