+++ api/core/v2/time_window.pb.go
@@ -973,10 +973,7 @@ func (m *TimeWindowWhen) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthTimeWindow
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthTimeWindow
 			}
 			if (iNdEx + skippy) > l {
@@ -1299,10 +1296,7 @@ func (m *TimeWindowDays) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthTimeWindow
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthTimeWindow
 			}
 			if (iNdEx + skippy) > l {
@@ -1417,10 +1411,7 @@ func (m *TimeWindowTimeRange) Unmarshal(dAtA []byte) error {
 			if err != nil {
 				return err
 			}
-			if skippy < 0 {
-				return ErrInvalidLengthTimeWindow
-			}
-			if (iNdEx + skippy) < 0 {
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
 				return ErrInvalidLengthTimeWindow
 			}
 			if (iNdEx + skippy) > l {